pyproject.toml
[tool.poetry]
name = "site"
version = "0.1.0"
description = "David Hariri's place on the web"
authors = ["David Hariri <david.hariri@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
gunicorn = "^22.0.0"
rfeed = "^1.1.1"
sentry-sdk = "^2.2.0"
markdown = "^3.5"
pydantic = "^2.4.2"
python-frontmatter = "^1.0.0"
python-dotenv = "^1.0.0"
pygments = "^2.17.2"
click = "^8.1.7"
pydantic-settings = "^2.5.2"
python-slugify = "^8.0.4"
openai = "^1.51.2"
boto3 = "^1.35.42"
tweepy = "^4.14.0"
beautifulsoup4 = "^4.12.3"
pymdown-extensions = "^10.12"
atproto = "^0.0.56"
quart = "^0.19.9"
motor = "^3.6.0"


[tool.poetry.group.dev.dependencies]
pytest = ">=8.2"
ruff = "^0.8.4"
pytest-asyncio = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
